DEMRUM-862: improve naming and code organization following codestyle doc
