+++ SplunkCustomTracking/Tests/SplunkCustomTrackingTests/Testing Support/XCTestCase+CustomTrackingDataAccessors.swift
@@ -1,0 +1,36 @@
+//
+/*
+Copyright 2025 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+@testable import SplunkCommon
+@testable import SplunkCustomTracking
+import XCTest
+
+extension XCTestCase {
+    func getStringValue(for key: String, in data: CustomTrackingData) -> String? {
+        if case let .string(value) = data.attributes[key] {
+            return value
+        }
+        return nil
+    }
+
+    func getIntValue(for key: String, in data: CustomTrackingData) -> Int? {
+        if case let .int(value) = data.attributes[key] {
+            return value
+        }
+        return nil
+    }
+}
